+++ src/main/java/kr/co/backend/repository/ContactRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContactRepository extends JpaRepository<Contact, Long>, CustomContactRepository {
+public interface ContactRepository extends JpaRepository<Contact, Integer>, CustomContactRepository {
 
     Page<Tuple> getAll(Pageable pageable);
